study_agent/requirements.txt
# requirements.txt
streamlit
pymupdf
google-generativeai
sentence-transformers
faiss-cpu
pandas
numpy
pytesseract
pillow
python-dotenv